Feature: created 'ModalItem' component, also separated components in different folders
