+++ src/components/Home/Home.tsx
@@ -1,12 +1,12 @@
 import React, { useState }  from 'react';
-import {Modal} from './Modal/Modal'
+import { Modal } from '../Modal/Modal'
 
 export const Home = (initialState = {}) => {
   const [values, setValues] = useState(initialState)
 
   const handleCloseModal = (e:any) => {
     setValues ({
-            //voy aqui
+      //voy aqui
     })
   }
 
@@ -15,9 +15,9 @@ export const Home = (initialState = {}) => {
     <section className="upload">
       <div className="upload__form">
         <input className="upload__files" type="file" />
-        <Modal isOpen={true} onClose={handleCloseModal}  />
+        <Modal isOpen={true} onClose={handleCloseModal}/>
       </div>
-      <p>Only jpg, png, gif, svg,</p>
+      <p>Only jpg, png, gif, svg</p>
     </section>
   )
 }
