+++ src/components/Modal/styles.scss
@@ -1,4 +1,4 @@
-@import '../base';
+@import '../../styles/base';
 
 .modal {
   position: fixed;
