+++ src/components/ModalItem/ModalItem.tsx
@@ -1,0 +1,10 @@
+import React, { useState } from 'react'
+import './styles.scss'
+
+export const ModalItem = (ImageUrl: any) => {
+  return (
+    <div className="modalItem">
+      <p>1</p>
+    </div>
+  )
+}
